Added profile page controller, model and route
