+++ controllers/PostController.js
@@ -53,67 +53,9 @@ const unlikePost = async (req, res) => {
   }
 };
 
-const getOwnTweets = async (req, res) => {
-  try {
-    const { userId } = req.query;
-    const query = await pool.query(postQueries.getOwnTweets, [userId]);
-    res.send(query.rows);
-  } catch (error) {
-    console.log(error);
-    res.status(500).send(error);
-  }
-};
-
-const getOwnReplies = async (req, res) => {
-  try {
-    const { userId } = req.query;
-    const query = await pool.query(postQueries.getOwnReplies, [userId]);
-    res.send(query.rows);
-  } catch (error) {
-    console.log(error);
-    res.status(500).send(error);
-  }
-};
-
-const getOwnLikes = async (req, res) => {
-  try {
-    const { userId } = req.query;
-    const query = await pool.query(postQueries.getOwnLikes, [userId]);
-    res.send(query.rows);
-  } catch (error) {
-    console.log(error);
-    res.status(500).send(error);
-  }
-};
-const getTweetCount = async (req, res) => {
-  try {
-    const { userId } = req.query;
-    const query = await pool.query(postQueries.getTweetCount, [userId]);
-    res.send(query.rows[0].count);
-  } catch (error) {
-    console.log(error);
-    res.status(500).send(error);
-  }
-};
-
-const getBio = async (req, res) => {
-  try {
-    const { userId } = req.query;
-    const query = await pool.query(postQueries.getBio, [userId]);
-    res.send(query.rows[0].Bio);
-  } catch (error) {
-    console.log(error);
-    res.status(500).send(error);
-  }
-};
 module.exports = {
   addPost,
   getPosts,
   likePost,
   unlikePost,
-  getOwnTweets,
-  getOwnReplies,
-  getOwnLikes,
-  getTweetCount,
-  getBio,
 };
