+++ index.js
@@ -8,6 +8,7 @@ const jwks = require("jwks-rsa");
 const pool = require("./database/db");
 const userRoute = require("./routes/userRoutes");
 const postRoute = require("./routes/postRoutes");
+const profileRoute = require("./routes/profileRoutes");
 
 const verifyJwt = jwt({
   secret: jwks.expressJwtSecret({
@@ -60,5 +61,6 @@ app.use(cors());
 
 app.use("/api/users", verifyJwt, currentUserCheck, userRoute);
 app.use("/api/posts", postRoute);
+app.use("/api/profile", profileRoute);
 
 app.listen(port, () => console.log(`Listening on port ${port}....`));
