+++ routes/postRoutes.js
@@ -8,10 +8,5 @@ router.get("/", postController.getPosts);
 router.post("/", postController.addPost);
 router.post("/likePost", postController.likePost);
 router.delete("/unlikePost", postController.unlikePost);
-router.get("/getOwnTweets", postController.getOwnTweets);
-router.get("/getOwnReplies", postController.getOwnReplies);
-router.get("/getOwnLikes", postController.getOwnLikes);
-router.get("/getTweetCount", postController.getTweetCount);
-router.get("/getBio", postController.getBio);
 
 module.exports = router;
